refactor(runtime): move session_repair inline tests into their own file
The 1,884-LOC `#[cfg(test)] mod tests { ... }` block at the bottom of
`session_repair.rs` moves into `session_repair/tests.rs`. The parent
file declares it with `#[cfg(test)] mod tests;` so the suite is
path-equivalent (`session_repair::tests::*` runs the same tests).

The dedent + de-nesting pulls the previously-inside-mod-tests `#[test]`
fns up one level so test names stay flat and the `super::*` import
resolves to session_repair (not session_repair::tests::tests).

session_repair.rs: 3,206 -> 1,318 LOC (-1,888). Now under the 2,000 LOC
file-size cap from #3710's acceptance criteria. The remaining file is
the actual repair pipeline: `validate_and_repair*`, the seven validator
phases (rescue_misplaced / enforce_adjacent / reorder / insert_synthetic /
deduplicate / remove_aborted / repair-helpers), the content-sanitizer
trio (`strip_tool_result_details` / `strip_base64_blobs` /
`strip_injection_markers`), and the trim helpers
(`prune_heartbeat_turns`, `find_safe_trim_point`, `is_safe_boundary`,
`coalesce_adjacent_text_blocks`).

cargo check --tests + cargo clippy --lib both clean.

Refs #3710
